fix(web): missing posthog env should not cause runtime crash (#648)
## Description

this pr makes sure that posthog not being set doesn't create a runtime error. (which is the current behavior)

## Testing
i tested mostly by changing the env vars inside of 

`.env.local` 

setting `POSTHOG_ENABLED` to true without a `POSTHOG_KEY` 
setting it to  false, etc.
and running `pnpm build:all` `pnpm start:all`

## Expected behavior
the expected behavior is that this should create a buildtime crash if `POSTHOG_ENABLED` is set to true with no `POSTHOG_KEY` set.


## What next?
i would've like to test it a bit more but the next logical step would be a preview like environment that would give me 99% confidence that all is working as expected. right now my confidence is more at 80%.

i think we should deploy this asap to staging, to verify that it works as expected (we would know if it doesn't crash since i've set the env vars in infiscal) to get confidence to 99%